refactor(db): split and moved transactional workflow update logic
